+++ CHANGELOG.md
@@ -14,3 +14,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - OAuth2 Security
 - OIDC <> Organisation Roles Mapping
 - Session User Info
+- Session User Info email